// src/app/@api/model/contactUsDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ContactUsDto { 
    firstName: string;
    lastName: string;
    email: string;
    phone?: number;
    companyName?: string;
    message: string;
}